move ScanningSpinner to common folder
