+++ src/common/components/scanning-spinner/scanning-spinner.tsx
