+++ src/tests/unit/tests/common/components/scanning-spinner/scanning-spinner.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { ScanningSpinner } from 'electron/views/automated-checks/components/scanning-spinner';
+import { ScanningSpinner } from 'common/components/scanning-spinner/scanning-spinner';
 import { shallow } from 'enzyme';
 import * as React from 'react';
